# generated/api/v1/collector_runtime_config_service.pb.gw.go
// Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
// source: api/v1/collector_runtime_config_service.proto

/*
Package v1 is a reverse proxy.

It translates gRPC into RESTful JSON APIs.
*/
package v1

import (
	"context"
	"io"
	"net/http"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/grpclog"
	"google.golang.org/grpc/metadata"
	"google.golang.org/grpc/status"
	"google.golang.org/protobuf/proto"
)

// Suppress "imported and not used" errors
var _ codes.Code
var _ io.Reader
var _ status.Status
var _ = runtime.String
var _ = utilities.NewDoubleArray
var _ = metadata.Join

func request_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0(ctx context.Context, marshaler runtime.Marshaler, client CollectorRuntimeConfigServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq Empty
	var metadata runtime.ServerMetadata

	msg, err := client.GetCollectorRuntimeConfig(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0(ctx context.Context, marshaler runtime.Marshaler, server CollectorRuntimeConfigServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq Empty
	var metadata runtime.ServerMetadata

	msg, err := server.GetCollectorRuntimeConfig(ctx, &protoReq)
	return msg, metadata, err

}

func request_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0(ctx context.Context, marshaler runtime.Marshaler, client CollectorRuntimeConfigServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq CollectorRuntimeConfigRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.PatchCollectorRuntimeConfig(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0(ctx context.Context, marshaler runtime.Marshaler, server CollectorRuntimeConfigServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq CollectorRuntimeConfigRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.PatchCollectorRuntimeConfig(ctx, &protoReq)
	return msg, metadata, err

}

func request_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0(ctx context.Context, marshaler runtime.Marshaler, client CollectorRuntimeConfigServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq CollectorRuntimeConfigRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.DryRunCollectorRuntimeConfig(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0(ctx context.Context, marshaler runtime.Marshaler, server CollectorRuntimeConfigServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq CollectorRuntimeConfigRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.DryRunCollectorRuntimeConfig(ctx, &protoReq)
	return msg, metadata, err

}

// RegisterCollectorRuntimeConfigServiceHandlerServer registers the http handlers for service CollectorRuntimeConfigService to "mux".
// UnaryRPC     :call CollectorRuntimeConfigServiceServer directly.
// StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterCollectorRuntimeConfigServiceHandlerFromEndpoint instead.
func RegisterCollectorRuntimeConfigServiceHandlerServer(ctx context.Context, mux *runtime.ServeMux, server CollectorRuntimeConfigServiceServer) error {

	mux.Handle("GET", pattern_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/v1.CollectorRuntimeConfigService/GetCollectorRuntimeConfig", runtime.WithHTTPPathPattern("/v1/collector_runtime_config"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("PATCH", pattern_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/v1.CollectorRuntimeConfigService/PatchCollectorRuntimeConfig", runtime.WithHTTPPathPattern("/v1/collector_runtime_config"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/v1.CollectorRuntimeConfigService/DryRunCollectorRuntimeConfig", runtime.WithHTTPPathPattern("/v1/collector_runtime_config/dryrun"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

// RegisterCollectorRuntimeConfigServiceHandlerFromEndpoint is same as RegisterCollectorRuntimeConfigServiceHandler but
// automatically dials to "endpoint" and closes the connection when "ctx" gets done.
func RegisterCollectorRuntimeConfigServiceHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
	conn, err := grpc.NewClient(endpoint, opts...)
	if err != nil {
		return err
	}
	defer func() {
		if err != nil {
			if cerr := conn.Close(); cerr != nil {
				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
			}
			return
		}
		go func() {
			<-ctx.Done()
			if cerr := conn.Close(); cerr != nil {
				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
			}
		}()
	}()

	return RegisterCollectorRuntimeConfigServiceHandler(ctx, mux, conn)
}

// RegisterCollectorRuntimeConfigServiceHandler registers the http handlers for service CollectorRuntimeConfigService to "mux".
// The handlers forward requests to the grpc endpoint over "conn".
func RegisterCollectorRuntimeConfigServiceHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
	return RegisterCollectorRuntimeConfigServiceHandlerClient(ctx, mux, NewCollectorRuntimeConfigServiceClient(conn))
}

// RegisterCollectorRuntimeConfigServiceHandlerClient registers the http handlers for service CollectorRuntimeConfigService
// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "CollectorRuntimeConfigServiceClient".
// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "CollectorRuntimeConfigServiceClient"
// doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
// "CollectorRuntimeConfigServiceClient" to call the correct interceptors.
func RegisterCollectorRuntimeConfigServiceHandlerClient(ctx context.Context, mux *runtime.ServeMux, client CollectorRuntimeConfigServiceClient) error {

	mux.Handle("GET", pattern_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/v1.CollectorRuntimeConfigService/GetCollectorRuntimeConfig", runtime.WithHTTPPathPattern("/v1/collector_runtime_config"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("PATCH", pattern_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/v1.CollectorRuntimeConfigService/PatchCollectorRuntimeConfig", runtime.WithHTTPPathPattern("/v1/collector_runtime_config"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/v1.CollectorRuntimeConfigService/DryRunCollectorRuntimeConfig", runtime.WithHTTPPathPattern("/v1/collector_runtime_config/dryrun"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

var (
	pattern_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "collector_runtime_config"}, ""))

	pattern_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "collector_runtime_config"}, ""))

	pattern_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2}, []string{"v1", "collector_runtime_config", "dryrun"}, ""))
)

var (
	forward_CollectorRuntimeConfigService_GetCollectorRuntimeConfig_0 = runtime.ForwardResponseMessage

	forward_CollectorRuntimeConfigService_PatchCollectorRuntimeConfig_0 = runtime.ForwardResponseMessage

	forward_CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig_0 = runtime.ForwardResponseMessage
)