# generated/api/v1/collector_runtime_config_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/collector_runtime_config_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CollectorRuntimeConfigService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/collector_runtime_config": {
      "get": {
        "operationId": "CollectorRuntimeConfigService_GetCollectorRuntimeConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetCollectorRuntimeConfigResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "CollectorRuntimeConfigService"
        ]
      },
      "patch": {
        "operationId": "CollectorRuntimeConfigService_PatchCollectorRuntimeConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Empty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CollectorRuntimeConfigRequest"
            }
          }
        ],
        "tags": [
          "CollectorRuntimeConfigService"
        ]
      }
    },
    "/v1/collector_runtime_config/dryrun": {
      "post": {
        "operationId": "CollectorRuntimeConfigService_DryRunCollectorRuntimeConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DryRunCollectorRuntimeConfigResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CollectorRuntimeConfigRequest"
            }
          }
        ],
        "tags": [
          "CollectorRuntimeConfigService"
        ]
      }
    }
  },
  "definitions": {
    "DryRunCollectorRuntimeConfigResponseStatusWithResources": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/commonapiStatus"
        },
        "resources": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1DryRunCollectorRuntimeConfigResponseResource"
          }
        }
      }
    },
    "DryRunCollectorRuntimeConfigResponseStatusesForFeature": {
      "type": "object",
      "properties": {
        "runtimeFeature": {
          "$ref": "#/definitions/commonapiRuntimeFeature"
        },
        "statusWithResources": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/DryRunCollectorRuntimeConfigResponseStatusWithResources"
          }
        }
      }
    },
    "RuntimeFeatureConfigRuntimeRule": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "configResourceCollection": {
          "$ref": "#/definitions/commonapiConfigResourceCollection"
        },
        "status": {
          "$ref": "#/definitions/commonapiStatus"
        }
      }
    },
    "StatusNetworkConnectionStatus": {
      "type": "object",
      "properties": {
        "enabled": {
          "type": "boolean"
        },
        "aggregateExternal": {
          "type": "boolean"
        }
      }
    },
    "StatusNetworkEndpointStatus": {
      "type": "object",
      "properties": {
        "enabled": {
          "type": "boolean"
        },
        "processesListeningOnPort": {
          "type": "boolean"
        }
      }
    },
    "StatusProcessStatus": {
      "type": "object",
      "properties": {
        "enabled": {
          "type": "boolean"
        }
      }
    },
    "commonapiBooleanOperator": {
      "type": "string",
      "enum": [
        "OR",
        "AND"
      ],
      "default": "OR"
    },
    "commonapiCollectorRuntimeConfig": {
      "type": "object",
      "properties": {
        "runtimeFeatureConfig": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonapiRuntimeFeatureConfig"
          }
        },
        "resourceCollections": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonapiResourceCollection"
          }
        }
      }
    },
    "commonapiConfigResourceCollection": {
      "type": "object",
      "properties": {
        "resourceCollectionName": {
          "type": "string"
        },
        "resourceCollection": {
          "$ref": "#/definitions/commonapiResourceCollection"
        }
      }
    },
    "commonapiMatchType": {
      "type": "string",
      "enum": [
        "EXACT",
        "REGEX"
      ],
      "default": "EXACT"
    },
    "commonapiResourceCollection": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "lastUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "createdBy": {
          "$ref": "#/definitions/commonapiSlimUser"
        },
        "updatedBy": {
          "$ref": "#/definitions/commonapiSlimUser"
        },
        "resourceSelectors": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonapiResourceSelector"
          },
          "description": "`resource_selectors` resolve as disjunction (OR) with each-other and with selectors from `embedded_collections`. For MVP, the size of resource_selectors will at most be 1 from UX standpoint."
        },
        "embeddedCollections": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonapiResourceCollectionEmbeddedResourceCollection"
          }
        }
      }
    },
    "commonapiResourceCollectionEmbeddedResourceCollection": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "commonapiResourceSelector": {
      "type": "object",
      "properties": {
        "rules": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonapiSelectorRule"
          },
          "description": "`rules` resolve as a conjunction (AND)."
        }
      }
    },
    "commonapiRuleValue": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        },
        "matchType": {
          "$ref": "#/definitions/commonapiMatchType"
        }
      }
    },
    "commonapiRuntimeFeature": {
      "type": "string",
      "enum": [
        "PROCESSES",
        "NETWORK_CONNECTIONS",
        "NETWORK_ENDPOINTS"
      ],
      "default": "PROCESSES"
    },
    "commonapiRuntimeFeatureConfig": {
      "type": "object",
      "properties": {
        "feature": {
          "$ref": "#/definitions/commonapiRuntimeFeature"
        },
        "defaultStatus": {
          "$ref": "#/definitions/commonapiStatus"
        },
        "rules": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimeFeatureConfigRuntimeRule"
          }
        }
      }
    },
    "commonapiSelectorRule": {
      "type": "object",
      "properties": {
        "fieldName": {
          "type": "string",
          "title": "`field_name` can be one of the following:\n- Cluster\n- Cluster Label\n- Namespace\n- Namespace Label\n- Namespace Annotation\n- Deployment\n- Deployment Label\n- Deployment Annotation"
        },
        "operator": {
          "$ref": "#/definitions/commonapiBooleanOperator",
          "title": "'operator' only supports disjunction (OR) currently"
        },
        "values": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonapiRuleValue"
          },
          "description": "`values` resolve as a conjunction (AND) or disjunction (OR) depending on operator. For MVP, only OR is supported from UX standpoint."
        }
      }
    },
    "commonapiSlimUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "commonapiStatus": {
      "type": "object",
      "properties": {
        "processStatus": {
          "$ref": "#/definitions/StatusProcessStatus"
        },
        "networkConnectionStatus": {
          "$ref": "#/definitions/StatusNetworkConnectionStatus"
        },
        "networkEndpointStatus": {
          "$ref": "#/definitions/StatusNetworkEndpointStatus"
        }
      }
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com. As of May 2023, there are no widely used type server\nimplementations and no plans to implement one.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(&foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n    // or ...\n    if (any.isSameTypeAs(Foo.getDefaultInstance())) {\n      foo = any.unpack(Foo.getDefaultInstance());\n    }\n\n Example 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\n Example 4: Pack and unpack a message in Go\n\n     foo := &pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := &pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\nJSON\n====\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": <string>,\n      \"lastName\": <string>\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "v1CollectorRuntimeConfigRequest": {
      "type": "object",
      "properties": {
        "collectorRuntimeConfig": {
          "$ref": "#/definitions/commonapiCollectorRuntimeConfig"
        }
      }
    },
    "v1DryRunCollectorRuntimeConfigResponse": {
      "type": "object",
      "properties": {
        "statusesForFeatures": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/DryRunCollectorRuntimeConfigResponseStatusesForFeature"
          }
        }
      }
    },
    "v1DryRunCollectorRuntimeConfigResponseResource": {
      "type": "object",
      "properties": {
        "cluster": {
          "type": "string"
        },
        "namespace": {
          "type": "string"
        }
      }
    },
    "v1Empty": {
      "type": "object"
    },
    "v1GetCollectorRuntimeConfigResponse": {
      "type": "object",
      "properties": {
        "collectorRuntimeConfig": {
          "$ref": "#/definitions/commonapiCollectorRuntimeConfig"
        }
      }
    }
  }
}